Added options to handle undeclared ids in .rlu files
The 2nd release of the Tokeneer case study involves package level RLU
(user-defined rules) files that are read when processing every unit in
the package, but that include rules involving constant or function
that are only declared in some of the units.

2 options are added to enable Victor to handle this situation:

- When processing a unit, delete any rules with identifiers that are
  undeclared in the FDL file for the unit

- When processing a unit, read in all FDL declaration files of units in 
  the same package.  

See the updated user manual for the command line switches for these
options and for the options for reading the user-defined rules files.