Changed handling of undeclared identifiers in rules
Until now, the default was that rules with undeclared function or
constant identifiers were left alone.  This caused type checking of
the program unit VCs to fail and no attempt would be made to prove the
VCs.  If the -delete-rules-with-undeclared-ids option were asserted,
then instead these rules would be deleted.

Now the behaviour is that rules with undeclared ids are always
deleted: this always seems to be a sensible course of action.
Therefore the -delete-rules-with-undeclared-ids option has been
removed.  

A warning message is now printed to the .vlg file about the deletion
unless the option -expect-user-rules-with-undeclared-ids is asserted.
(Undeclared Ids sometimes occur in the user rules file for a directory
of program units.  We never expect the Examiner to generate rules with
undeclared ids.)