Added user rule auditing feature
This feature helps with checking if any user rules are inconsistent,
and with finding whether user rules are provable with or without the
use of other user rules.  This feature complements the feature for
finding redundant rules.


See Section 4.3 "Analysing user rules" of the User Manual for
documentation on both features.

Added run/Makefile option variables to control the auditing.