Improved handling of integer exponent function
Until now, Victor's default behaviour on encountering terms of form
"n ** m", with n,m natural numbers, has been to evaluate them, as SMT
solvers do not provide support for the exponent function.

However, certain rules in Victor's prelude, borrowed from the rule set used
by the SPARK checker tool, for example

bitwise(67):  bit__xor(X,Y) <= 2**N - 1  may_be_deduced_from 
                                         [ 0 <= X, 0 <= Y, 0 <= N,
                                           X <= 2**N - 1, Y <= 2**N - 1].

can only be successfully instantiated by SMT solvers when a binding
for N can be inferred from the discovery of expressions involving **.

One place this rule is used is in verifying VCs from the SPARKskein programme
in the SPARK library.

The previous work-around to this problem was that the user had to
disable evaluation of ** when using Victor alone for verifying VCs for
this programme, and rely on rules giving values of particular
instances of ** in the Victor prelude rules file.  However evaluation
still needed to be enabled for other verifications which did involve
ground terms of form "n ** m".

A better solution has now been implemented which should work by
default in most cases.  This involves using a new option
-add-ground-exp-eval-axioms, instead of the option -ground-eval-exp.
This new option adds a new rule

   n**m = eval(n**m)
  
where n,m are natural numbers and "eval" means evaluate the argument
to a natural number, if

 1. n**m is found in unit, or
 2. natural number eval(2**m) or eval(2**m)-1 
      for some m in range 1..option("exp-2-bound").  (64 default upper bound)
    is found in unit.

The run/Makefile-prelude.mk has been modified to use this new option
instead of -ground-eval-exp, and the instances of ** in the
run/prelude.rul file have been removed, as they are no longer
necessary.

See the User Manual for documentation on the new option.